requirements.txt
langchain
langsmith
qdrant-client
sentence-transformers
torch
pypdf
streamlit
requests
python-dotenv
langchain_community
langchain_core
langchain-groq